Create hashpasswordlib.py
A mathematical algo setup in a py code. What it does is simple, get a randomized hash of a string, however, with each hashing of the same string, you can see, the hashed strings are not the same, rather randomized. 

Just created this file, would obviously be tweaking a lot more into it. Really a lot.